move WAB::Configuration into WAB::Impl namespace
