+++ lib/wab.rb
@@ -4,7 +4,6 @@ module WAB
 end
 
 require 'wab/controller'
-require 'wab/configuration'
 require 'wab/data'
 require 'wab/errors'
 require 'wab/shell_logger'
